bur_controller refactoring
